Week6/requirements.txt
pandas
numpy<2
scikit-learn
streamlit
plotly
joblib
tensorflow-macos
tensorflow-metal
